server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.21.0
mrtparse>=1.6.0
pymilvus>=2.3.4
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
jinja2>=3.1.0
pydantic>=2.0.0
tqdm>=4.65.0
requests>=2.31.0
loguru>=0.7.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pybgpstream>=2.0.0

# MCP / LangGraph
fastmcp
requests
langchain
langchain-openai
langchain-mcp-adapters
langgraph

# OpenAI evals (GitHub에서 직접 설치)
git+https://github.com/openai/evals.git